+++ apps/web/src/studio/features/analytics/agent/components/AgentList.tsx
@@ -1,5 +1,7 @@
+import { Button } from "@caseai-connect/ui/shad/button"
+import { Settings2Icon } from "lucide-react"
 import { useTranslation } from "react-i18next"
-import { useOutlet } from "react-router-dom"
+import { Link, useOutlet } from "react-router-dom"
 import { Grid, GridContent, GridHeader } from "@/common/components/grid/Grid"
 import { selectAgentsData } from "@/common/features/agents/agents.selectors"
 import { AgentItem } from "@/common/features/agents/components/AgentItem"
@@ -11,8 +13,8 @@ import { DocumentsButton } from "@/studio/features/agents/components/DocumentsBu
 import { EvaluationButton } from "@/studio/features/agents/components/EvaluationButton"
 import { MembersButton } from "@/studio/features/agents/components/MembersButton"
 import { ProjectDeletor } from "@/studio/features/projects/components/ProjectDeletor"
-import { ProjectEditor } from "@/studio/features/projects/components/ProjectEditor"
 import { ReviewCampaignsButton } from "@/studio/features/review-campaigns/components/ReviewCampaignsButton"
+import { StudioRoutes } from "@/studio/routes/helpers"
 
 const extraItems = [
   AgentCreatorButton,
@@ -37,7 +39,17 @@ export function AgentList() {
         description={t("project:project")}
         action={
           <>
-            <ProjectEditor project={project} />
+            <Button variant="outline" asChild>
+              <Link
+                to={StudioRoutes.projectAdmin.build({
+                  organizationId: project.organizationId,
+                  projectId: project.id,
+                })}
+              >
+                <Settings2Icon />
+                {t("actions:edit")}
+              </Link>
+            </Button>
             <ProjectDeletor project={project} />
           </>
         }